QR Code/requirement.txt
qrcode[pil]
pyqrcode
pypng
segno
